instagram-scraper/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
lxml==4.9.3
pillow==10.2.0
pandas==2.1.4
tqdm==4.66.1
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
